feat: implement GetSchemaString for MySQL
